fix(config): enforce web FeatureType parity with canonical CONFIG_FEATURE_TYPES (#2004)
The canonical config-policy feature-type list (CONFIG_FEATURE_TYPES) lived in
apps/api and the web layer re-declared it by hand in two places, so a new
canonical type would silently fail to render a card/tab until each web list was
edited by hand (a recurring, tsc-clean omission — #2003 fixed one instance).

Make @breeze/shared the single source of truth and derive the web unions from
it (issue option 1), backed by runtime parity tests (option 2):

- Move CONFIG_FEATURE_TYPES / ConfigFeatureType into
  packages/shared/src/constants/configFeatureTypes.ts; apps/api re-exports it
  (the existing Drizzle-enum parity test in policyBaselineDefaults.test.ts still
  pins it to the DB enum through the re-export).
- Web feature-type unions now derive via Exclude<ConfigFeatureType, …>:
  - featureTabs/types.ts: minus 'onedrive_helper' (no editor tab).
  - DeviceEffectiveConfigTab.tsx: minus 'remote_access' | 'pam' (applied-when-
    unassigned baselines). Their exhaustive FEATURE_META Records make a new
    canonical type fail tsc until accounted for.
- Add web parity tests asserting each surface equals canonical minus its
  documented exclusions, and that the exclusions are real canonical members.

Verified non-vacuous: adding a fake canonical type fails both tsc (the two
FEATURE_META Records) and all three parity tests; reverting makes them green.
