+++ apps/api/src/services/configFeatureTypes.ts
@@ -1,25 +1,22 @@
 /**
- * Canonical list of configuration-policy feature types.
+ * Canonical configuration-policy feature types — api-side entry point.
  *
- * This lives in its own LEAF module (no heavy imports) on purpose. Light
- * consumers — `policyBaselineDefaults.ts`, and transitively `pamSettings.ts`
- * and the agent `helpers.ts` — need the feature-type list without dragging in
- * the full `configurationPolicy.ts` service, which references DB schema tables
- * at module-eval (e.g. FEATURE_TABLE_MAP). Importing the list from here instead
- * of from `configurationPolicy.ts` breaks the configurationPolicy ⇄
- * policyBaselineDefaults import cycle and keeps route/helper test suites (which
- * use partial `db/schema` mocks) from crash-loading the service. (#1725)
+ * The list itself now lives in `@breeze/shared` (constants/configFeatureTypes.ts)
+ * so the web layer can derive its feature-type unions from the same source and
+ * the two can never silently drift (#2004). This module re-exports it; it stays
+ * a LEAF re-export (no heavy imports) on purpose — light consumers
+ * (`policyBaselineDefaults.ts`, and transitively `pamSettings.ts` and the agent
+ * `helpers.ts`) need the feature-type list without dragging in the full
+ * `configurationPolicy.ts` service, which references DB schema tables at
+ * module-eval (e.g. FEATURE_TABLE_MAP). Importing from here instead of from
+ * `configurationPolicy.ts` breaks the configurationPolicy ⇄ policyBaselineDefaults
+ * import cycle and keeps route/helper test suites (which use partial `db/schema`
+ * mocks) from crash-loading the service. (#1725)
  *
- * `configurationPolicy.ts` re-exports both names, so existing importers that
- * pull `ConfigFeatureType` / `CONFIG_FEATURE_TYPES` from there keep working.
+ * `configurationPolicy.ts` re-exports both names, so existing importers that pull
+ * `ConfigFeatureType` / `CONFIG_FEATURE_TYPES` from there keep working.
  *
- * A parity test asserts this list matches the Drizzle `configFeatureTypeEnum`.
+ * A parity test (`policyBaselineDefaults.test.ts`) asserts this list matches the
+ * Drizzle `configFeatureTypeEnum`.
  */
-export const CONFIG_FEATURE_TYPES = [
-  'patch', 'alert_rule', 'backup', 'security', 'monitoring', 'maintenance',
-  'compliance', 'automation', 'event_log', 'software_policy', 'sensitive_data',
-  'peripheral_control', 'warranty', 'helper', 'remote_access', 'pam', 'onedrive_helper',
-  'vulnerability',
-] as const;
-
-export type ConfigFeatureType = typeof CONFIG_FEATURE_TYPES[number];
+export { CONFIG_FEATURE_TYPES, type ConfigFeatureType } from '@breeze/shared';
