+++ apps/web/src/components/configurationPolicies/featureTabs/types.ts
@@ -1,4 +1,14 @@
-export type FeatureType = 'patch' | 'alert_rule' | 'backup' | 'security' | 'monitoring' | 'maintenance' | 'compliance' | 'automation' | 'event_log' | 'software_policy' | 'sensitive_data' | 'peripheral_control' | 'warranty' | 'helper' | 'remote_access' | 'pam' | 'vulnerability';
+import type { ConfigFeatureType } from '@breeze/shared';
+
+// Derived from the canonical CONFIG_FEATURE_TYPES (single source of truth in
+// @breeze/shared) so the config-policy editor's feature tabs can't silently
+// drift from the registry. `onedrive_helper` is deliberately excluded: it has
+// no editor tab (the OneDrive Helper is configured elsewhere). Deriving via
+// Exclude means a new canonical feature type makes FEATURE_META below fail to
+// compile until a tab entry is added, and featureTypeParity.test.ts asserts the
+// exclusion stays honest. (#2004)
+export type FeatureType = Exclude<ConfigFeatureType, 'onedrive_helper'>;
+export const EDITOR_EXCLUDED_FEATURE_TYPES = ['onedrive_helper'] as const;
 
 export type FeatureLink = {
   id: string;
